requirements.txt
annotated-types==0.7.0
anthropic==0.60.0
anyio==3.7.1
asgiref==3.11.0
attrs==25.3.0
bashlex==0.18
certifi==2025.4.26
charset-normalizer==3.4.3
distro==1.9.0
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docker==7.1.0
dotenv==0.9.9
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
libtmux==0.46.2
numpy==2.2.6
openai==1.1.1
outcome==1.3.0.post0
packaging==26.0
Pillow==10.1.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.5
selenium==4.15.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wsproto==1.2.0
